(PC-17806) refactor(auth): extract and move all mutations in api folder
